dc-and-mcp/pyproject.toml
[project]
name = "dc-and-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.50.0",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
